general fixes for rendering components
