+++ src/components/buttons/RadioButton/RadioButton.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import React from "react";
 import styles from "./radioButton.module.css";
