+++ src/components/buttons/RadioButton/RadioButtonGroup.tsx
@@ -1,8 +1,6 @@
-"use client";
-
 import React from 'react';
 import styles from "./radioButton.module.css";
-import { RenderOptions } from './RenderOptions';
+import { RadioButton } from './RadioButton';
 
 interface IOption {
     id: string;
@@ -14,22 +12,39 @@ interface RadioButtonGroupProps {
     id: string;
     label: string;
     options: IOption[];
-    onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+    selectedValue: string;
+    onValueChange: (name: string, value: string) => void;
 }
 export const RadioButtonGroup: React.FC<RadioButtonGroupProps> = ({
     id,
     label, 
     options,
-    onChange
+    selectedValue, 
+    onValueChange
 }) => {
+
+    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        onValueChange(e.target.name, e.target.value);
+    };
+
     return (
         <fieldset className={styles.fieldset} id={id}>
             <legend className={styles.legend}>{label}</legend>
             <div className={styles.wrapper}> 
-            <RenderOptions options={options} onChange={onChange} />
-                {/* <RadioButton key="radio-one" label="Bachelor" id="radio-one" name="radio" disabled={false}/> 
-                <RadioButton key="radio-two" label="Master" id="radio-two" name="radio" disabled={false}/>  */}
-            </div>
+            {options.map(({ id, label, disabled }, index) => (
+                <RadioButton
+                    key={id}
+                    id={id}
+                    label={label}
+                    name="radio"
+                    disabled={disabled}
+                    value={label}
+                    checked={selectedValue === label}
+                    defaultChecked={index === 0}
+                    onChange={handleChange}
+                />
+            ))}
+            </div> 
         </fieldset>  
     );
 };
